Move Metdata & Axis struct to vds.go
As well as being usefull for testing, these are also usefull for
documentation generation. Because they are used for testing the return
value from CGO (which ultimatly becomes the response body), we can safely
use them for documentation generation. Even tho they are never actually
used in the request handlers. Any out-of-sync issues when changing the
json format will be captured by tests.