Remove unwanted query decoding
In get requests we expect query parameter to be fully URL-encoded before
it is sent. gin should decode everything for us by default.
By decoding it one more time we are in danger of decoding too much, for
example from signature in SAS token which is expected to be send further
in its original form.