Change return Content-Type to multipart/mixed
multipart/mixed Content-Type[1] is common enough to have support in many
languages.

Returning multipart content solves two problems:
1. We are able to return metadata together with the result, without need
to query an additional endpoint.
2. When user request is processed successfully, but results in error due
to original file issues or incompatibility with provided data, we need
to return user information about it. However as it is not a 400-error
(user did provide correct data) and not a 500-error (server did process
the request, it just that result is not per user liking), the best
option would be to return 200. But for this user needs a way to
distinguish between valid requested data and response that contains an
error.
Ability to provide common metadata in multipart response gives us this
opportunity.

Thus this commit updates slice and fence endpoints to return
multipart/mixed data with 3 parts:
1) "application/json" common stream with information about occurred
errors.
2) "application/json" metadata relevant for the request. It contains
data from previous slice/metadata endpoint for slice and is empty for
fence.
3) "application/octet-stream" with previously returned data.

[1]
https://www.iana.org/assignments/media-types/media-types.xhtml#multipart