Add memory-leak check tests
We are not sure if these tests can bring any benefit or would be total
inconvenience.

We want to make sure that our (however small) cpp code doesn't leak
regardless of any possible future updates and refactoring. We also know
that some of our dependencies leak in certain cases, and having
memory-leaking code checked out can make it easier to have overview over
that behavior.

On the other side, these tests might be difficult to maintain, as they
kind of deal with "private" code (as we do not consider c++ code to be
our public interface). Unfortunately we don't know a way to assure
absence of memory leaks in our server through public code.

So for now we decided to introduce them, see how it goes in the future
and remove them if they are trouble and not just a dead weight.