Add metadata endpoint
The metadata endpoint returns volumetric data about a specific VDS. I.e.
axis descriptions. And data type. This should give enough information
for clients to construct sensible queries to the other endpoints for
now, but should be enriched with more metadata in the future. For now we
can add more entries on a per-need basis.