Reflow api/query.go
This file mostly groups together the enpoint-handlers (methods) at the
end of the file, while keeping helpers and internal methods further up.
Except for the endpoint.fence method that have snuck itself in at the
bottom together with the public endpoint methods. This commit move this
method up where the rest of it's kind resides.