Rename FenceRequest.Fence -> Coordinates
Struct field should be named the same as their json tag. If the
unmarshalling of the request fails on the required tag, the error
message will print the struct field name, not the json name. When the
names differ you'd get a missleading error message. In this case it
would say that the request is missing a required field "Fence", while in
reality what it's missing is "coordinates".