Add gzip support to endpoints
This commits add support for clients to ask for their response to be
gzip'd by specifying 'gzip' as part of the 'Accept-Encoding' header.

At the moment we priorities compression performance over compression
ratio. Compressing to aggressivly will have a negative impact on
cloud-to-cloud. This might change if we do more benchmarking.

The current compression ratio is about 2 for a farily dense binary
fence.