Refactor the Endpoint setup
The main motivation for this commit is to better facilitate future
support for multiple store-accounts in the same running instance.

One design change is to seperate the creation of vds-connection into
different functions. Particularly creation of vds connections based on
the file protocol, which is only used in test, are now situation solely
in the test-code.