Pin Alpine version of Docker container
The Go container used by the Dockerfile has been updated to use a new
release of Alpine Linux (3.17). This breaks the installation procedure
of OpenVDS within the container as OpenSSL v3 instead of v1 is used
within the new release. Thus we pin the Alpine version to 3.16 which
uses OpenSSL v1.