Swap axis order in slice metadata
The axis information does not match the way the data is ordered. One can
argue that it's the data that should be transposed, not the axis as that
seams to be the most natural orientation for e.g. plotting. However, I'm
unsure if we should transpose data server side as that imposed
additional compute on us.

Hence this commit reorders the metadata rather than the data, to make
sure they correspond. This descision might be revisited before we do a
proper versioned release of the API.