Make vds.Connection an interface
This commit converts the vds.Connection struct into an interface and
re-introduces the the old Connection struct as AzureConnection. Which
implements the new Connection interface.

The AzureConnection behaves exactly as the old Connection but the
the Url and Connections are now methods, not members.

The FileConnection is implemented to support test cases.

This change is motivated by some planned extensions to the Connection /
AzureConnection.