Extend the Connection interface with HasReadAccess
When implementing a response cache we need a way to validate the
connection if we get a cache hit. This commit extends the connection
interface to include a method HasReadAccess for exactly this purpose.

For AzureConnection this means doing a HEAD request to validate the sas
token, together with a manual check of the srt field.

For FileConnections there is always read access.