Rename vdsbuffer to requestdata
Making the naming for the request's data more generic. The buffer and
its content should be (ideally) dependent on the API and the API's
request, but not on the file format of the data accessed. We stick with
lower case names for the struct as this is common in C and this struct
is used in the C interface to Go.