Manually validate the srt field in sas-tokens
OpenVDS segfault on sas token where the srt field (allowed resource
types) includes 'c' (container), but not 'o' (object/blob). This
effectlivly takes down our server. Such tokens are valid in a general
sense, but does not include enough access to read the OpenVDS.

This commit implements a preventive verification of the srt field which
makes sure such tokens are never passed to OpenVDS. The request will be
catagorized as a bad request and aborted.