Add base class for handling seismic VDS data
This adds a base class SeismicHandle for handling data in a VDS
dataset. It is intended to be extended by corresponding classes for
different types as seismic data, e.g., pre-stack and post-stack data.
The commit also adds This adds a class which validates the base
assumptions for VDS data sets. The validator is called during the
creation of a SeismicHandle object. At the moment it only checks if
the dataset is three dimensional.