Don't check for 'srt' on Service & User Delegation Sas
This commit makes sure that validateSasSrt does not return an error if
'srt' is missing. This behaviour was never intended and returning an
error is considered a bug as both Service- and User Delegation SAS
doesn't use this parameters at all. As a consequence requests with such
SAS-tokens would fail.