Add validators for poststack data
This adds two validator: One validator specific to poststack data and
one validator specific to slice requests on poststack data.

The poststack data validator checks for the correct axis order in the
VDS. The slice request validator check whether the axis name and axis
units are sensible.

Valid names and units of axis are now stored in an  unordered_map of
lists of strings which nicely groups the valid axis name with valid axis
units. It allows for simple checks whether a unit or name is contained
in the valid units/names using STL functions. In order to do the
validations using the AxisDescriptor a new function has been added to
the AxisDescriptor which returns the name of an axis corresponding to
our API/request instead of the internal name used by VDS.