Add AxisDescriptor class
This introduces a AxisDescriptor class which extends the class
AxisMetadata. The class knows what coordinate system it belongs to,
what its name is in OpenVDS' notation, and what its axis dimension is.
This groups commonly accessed information about axes in a compact way.
This class will be used more in future commits when more functionality
is moved into classes that can access the seismic handle. Without access
to this handle one cannot create meaningful descriptors.