Move line number conversions to seismic handle
During a request the supplied line number has to be converted to VDS'
voxel coordinate system. The conversion depends on the request's
coordinate system (index or annotation). Since the seismic handle has
information about the use coordinate system and the mapping, the line
number conversion was moved into the seismic handle.