Add caching on slice and fence endpoints
This commit introduces in-memory cache of responses from the slice and
fence endpoints. The cache size (maximum memory footprint of the cache)
is customizable on server startup, and can be set to zero to disable
caching all together.

Cache package
-------------

The new cache package introduces a cache interface and two concrete
implementations. The first is an abstraction over a 3rd party cache
library called Ristretto which implements a LRU cache that is limited by
memory-usage. The other is a no-cache, i.e. a dumb implementation that
never caches and always result in cache misses. This is useful if we
want to disable caching all together. The construction of new caches is
unified in by the function NewCache(cachesize) which dispatches to the
right cache-implementation based on cachesize.

Constructing cache keys
-----------------------

Cache keys are constructed by hashing the parts of the request structs
(SliceRequest or FenceRequest) that will contribute towards a unique
response. I.e. every field except the sas token. E.g. two requests for
the same slice, but with different (valid) sas-tokens will still result in a
cache hit.

Authentication
--------------

Cache hits needs to be autheticated before the result can be returned.
This is implemented by doing a HEAD request on the VolumeDataLayout blob
and doing some extra checks on the sas token.