Cleanup metadata fetching for axes
This commit removes methods for fetching metadata of axes from the
post-stack handle as well as overly specific methods for metadata
fetching. Now, one can request the metadata of all axis from the
SeismicHandle as this class has all information needed for obtaining
axis metadata. It is then the responsibility of the caller to use the
metadata fields one needs.

The API accesses/stored metadata information in different order
depending on the metadata request. Due to that, there is the need to
reverse the order of axis metadata obtained for some cases.