Adjust E2E tests for internal Azure changes
Unexpectedly an Azure test that is in the code just as a very simple
security check started failing.
Apparently Azure now may return different error messages based on
whether 'x-ms-version' header was supplied with the request or not.

As it is an internal logic change, allow a set of error messages.
The point of this test anyway is just to assure that some kind of
"authentication failed" message is returned for wrong sas.