Include thresholds into metric name
I can't find a good way to print expected thresholds in k6. Not having
expected values in the list is quite inconvenient, as one has to check
source to know why it failed.

It probably can be done better when creating own text summary, but it is
more work for something we don't need that much. Adding thresholds to
metrics name should be enough.