Log requests
Log only successfully parsed requests. Goal of Requests logging is to
assist in future debugging of server behavior (which we got burned on
recently).
For malformed requests enough information is returned to the user. So it
is their responsibility to fix the request. Us logging it won't help
anything.

While information can be tied to the user's ip, it is not stored
anywhere for further processing.
SAS are not logged. Any other relevant information (aka vds) while isn't
public, shouldn't do any harm if known to anyone with access to the
server logs.