Rename axis enumeration to api_axis_name
In order to allow the name "axis" to be used for a class representing
axes in the coordinate system the enumerate is renamed tom to
`api_axis_name`. The new name also reflects that the enumeration
represents the name an axis may have in the API and thus as part of a
request. The name may not be present in the actual VDS dataset.