Run /metrics endpoint on different port than server
While we want to share our great server with the world by exposing it on
the internet, we don't want the world to read our server statistics. By
hosting /metrics on a different port we can keep it private.

Gin doesn't allow running routes on different ports out-of-the box.
This means that we need to run two gin instances. One for the server and
one for the /metrics (if turned on). This has the added benefit of not
poluting the server logs with requests to the /metrics endpoint. Which
is intended to be scraped continually.