Add Axis class
The class wraps around the OpenVDS axis descriptor and adds an
additional field for the dimension of the corresponding axis. The
dimension refers to the dimension/index of the current index in the VDS
voxel coordinate system and is needed for accessing data from a VDS
dataset.

The class simplifies and unifies the handling of information about axes.
This allowed to remove some functions that mapped between the different
notions of coordinate systems for requests and VDS' voxel coordinate
system.

As part of the changes the function `axis_todim` was changed into
`make_axis` which returns the corresponding Axis for a given `axis_name`
instead of the axis' dimension in the coordinate system of a request.