Add wrapper around OpenVDS' access manager
The DataHandle is essentially a wrapper around OpenVDS' access manager
with a simplified interface for our use case. Where MetadataHandle wraps
around OpenVDS' layout the DataHandle wraps around OpenVDS' access
manager.

The DataHandle constructs and maintains an instance of MetadataHandle
which is accessible through its interface. This means that the
DataHandle is the main entrypoint into our interfaces for the caller.