Define trace type as single array
The previous definition for the traces type consisted of a list of
traces without being able to know the actual number of traces. Changing
the trace type to represent a single trace instead of an array of traces
allows a more expressive construction of the traces array. This also
allows to construct a list of traces without any knowledge about how the
trace type is defined.