Move Metadata.format function to vds.cpp
When we talk about format, we might have 3 things in mind now:
 1. format in which data is stored in vds
 2. format in which we request data as
 3. format string we return to the user

Name Metadata.format seems most likely refer to 1, when in reality it
refers to 3. As this is a bit confusing, function is moved to vds.cpp
and renamed to fmtstr to make its purpose a bit more obvious.