Update staticcheck 0.3.3 -> 2023.1.2
The old version of the static analyzer does not support Go 1.20. Support
for Go 1.20 seems to be mandatory since we upgrade the Docker base
image. The versioning scheme of staticcheck has changed and therefore
also the donwload link has changed. In order to update the version in a
simpler way than before the version number is defined as argument.