Add metadata to horizon endpoint
To be perfectly honest, this round trip to C++ seams utterly pointless.
The only relevant metadata to return is the shape, which is identical to
the inputed horizon shape, and the format, which is always f4.

This is true for fence too and it is therefore implemented the same way
for consistency. However, we should seriously consider short-circuiting
these methods in Go. Both for performance and for code bloat.