Add core horizon support
This commit implements a handler for reading seismic along an horizon.
The handler more or less follows xtgeo's surface definition [1] which is
a 2D grid acompanied by an origin point in coordinates, physical
distances between rows and cols (xinc, yinc) and a rotation.

Internally we map the coordinate information to an affine matrix
(rotation/translation matrix), which is easier to work with when
converting from map indicies to world coordinates.

The fillValue is used to indicate "dead" samples in the horizon, and
will be written back in the output map. I.e. fillValue in -> fillValue
out. It's also used for out-of-bounds samples in the horizontal plane

Boundschecking
---------------

1) Any sample that is out-of-bounds in the vertical plane is an
   error.

2) Any sample that is out-of-bounds in the horizontal plane is filled
   with the fillvalue in the resulting map. This allows regional
   horizons to be used on production volumes.

[1] https://xtgeo.readthedocs.io/en/stable/datamodels.html#surface-regularsurface