Do not explicitly test for 400 and 500 in E2E
Original thought was to make one test for both kinds of errors -
BadRequest and InternalServerError. But as from server point of view
there is no difference in what type of error it logs, we can reduce
number of error e2e tests, check for just any error and trust that it is
properly tested in go as now go tests use almost same setup as server.