Move coords from upper-left to center in horizon
OpenVDS' transformers and OpenVDS data request functions have different
definitions of where a datapoint is. E.g. A transformer (To voxel or ijk) will
return (0,0,0) for the first sample in the cube. The request functions on the
other hand assumes the data is located in the center of a voxel. I.e. that the
first sample is at (0.5, 0.5, 0.5). This is a *VERY* sharp edge in the OpenVDS
API and borders on a bug. It means we cannot directly use the output from the
transformers as input to the request functions.

For horizon request we did already ajust, but in a very implicit way
without any documentation. This commits makes it explicit and documented
behaviour.

Intrestingly, we already noticed this for nearest interpolation and we
had measures in place for shifting voxels for that particular case.
Those are now redundant as we fixed the root case.