Add attribute samplevalue
The samplevalue attribute return the exact same map as we provide
through the /horizon endpoint today. In the future that endpoint will be
removed in favor of this attribute.

In order to extract the right index from the vertical windows we rely
on a computation depending on the vds samplerate. This is information we
don't have access to in Go. Hence we have to pass in the credentials and
re-open the vds in the C.attribute call. This is very unefficient, and a
testemony of the limitations in the current C API design. In the future
we will completly remove the seperate horizon endpoint and then we can
merge C.horizon and C.attribute, which will remove this inefficiency
again.