Add support for subsampling in /attribute
This commit lets the user specify a samplerate for their target vertical
window. For any practical purposes that means subsampling. All the noise
and test in this PR comes from the fact that we interpolate in the
vertical window our self, instead of letting OpenVDS do so. The
motivation for that is performance.

On my machine a request with a surface of size 400 x 700 and a vertical
window size of 401 (above = below = 20ms, samplerate = 0.1) took 29 sec
to complete (fetch + calculation). Doing it ourself takes that down to 5
sec. Doing calculation concurrently takes it down further, to below 2
sec for fetch + calculation. Concurrency is not part of this commit
though.

When interpolating our self we want to make sure that openvds does not
interpolate in the vertical plane. I.e. we must take to translate the
requested vertical windows to a set of vertical points that are at voxel
centers in the vds. Once that window is read, we have to shift it back
and interpolate it up to the target samplerate.