Compute attributes concurrently
Attribute calculations are very parallelizable as vertical windows are
independent of each other. Too paralellize all we need to do is to
manage the positions in the input buffer (horizon) and the output
buffers (attributes).

The concurrency is being diven by go and goroutines.

There is more work to be done to find the optimal concurrency level, and
perhaps make it customizable to the server admin.