Expose the vds' vertical samplerate in Go
Having C.attribute opening a new vds handle to read the vertical
samplerate will be suboptimal when we add support for concurrent
attribute calculation through Go. Then each goroutine would have to open
it's own handle just to read this one float. Thus this pre-read of
samplevalue.

This commit show very well that our C API is too simplistic and would
benefit from a major upgrade. Firstly, every call to C is stateless and
must re-open a vds-handle - which is highly inefficient. Secondly, we
have no good way of returning errors, other than through the response
struct. However that struct is unfit for simple data like this
samplerate as it would require us manually parse out the float from the
buffer in Go, which is just ugly and error prune as it kills all
typesafety.

Redesigning the C API is for later of course, and for know we have to
live with this uglyness.