Remove e2e tests for horizon
Horizon is soon to be removed in favor of attributes. Ideally we just
port these tests to call the attributes endpoint instead, but they all
rely on the well_know_default.vds, which unfortually will soon be
incompatible with the attributes endpoint as it has too short traces for
the interpolation method that will come. Hence they are temporarily
removed