Introduce a context object in the C API
The context is not yet used, but the intended usecase is to let all C
function accept it as a first argument, and use it to write error
messages into should the function fail. The caller can then read the
error message out of the context when the the function call fails. This
will remove the need for appending err strings to the response struct.

In theory it's usecase can be extended passed error messages if we find
a need for it elsewhere.