Signal errors through return value in C API
For now only a crude set of status codes are implemented, and no special
handling of the different ones are implemented in Go. This is future
work which will enable us to give the callers better errors when
something goes wrong in C/C++.

The error message itself is still returned via the response struct, but
that is subject to change as well.