Move error strings from response to context
This commit adds a Context* as the first parameter to all C functions.
When a function errors and return a non zero error code, an error
message is written to the context. Which in turn can be obtained from Go
using C.errmsg(ctx).

This removes the need to store errmsg on the response struct, which can
now be a pure buffer protocol.