Move OpenVDS::Open out of DataHandle constructor
Moving OpenVDS::Open out of the constructer is now feasable because
opening a new DataHandle is centralized to a single place in the code.
The main motivation for this is to get around the hack of wrapping the
metadata member in a std::unique_ptr. This was previously needed due to
MetadataHandle's missing default constructor and delayed initialization
rules, which didn't play nicely together. Now the metadata member can be
moved to the initializer list and the hack goes away.