Implement vds.Get* as methods on vds.VDSHandle
By moving all these functions to methods on the VDSHandle we can re-use an open
handle for multiple request. That will gain us some speed on requests that up
until now opened the vds file twice (and hence download the vds json docs
twice).

We also re-use the handle's ctx and together with the reuse of open handles
this removes quite a bit of code bloat.