Centralize mapping from C status to Go error
Now that we can propegate different error types in C++ back to Go, we
can start mapping these to Go errors and ultimatly to HTTP errors.
Although that is not yet implemented, this commit makes sure that such
mapping can happen in a single location, rather than in every Go
function that does a CGO call.