Expose OpenVDS handles to go
Opening new VDSHandles is not free, as every new open would eventually
trigger a new download of the LayerStatus and VolumeDataLayout. So first
and foremost, this paves the way for a performance optimization in fence
and slice where we currently open two handles for the VDS file, one for
the metadata call and one of the data call.

But by preserving the VDSHandle in Go, and thus making round-trips to
C++ cheaper we open the door for more optimizations and other
improvements as well.

Firstly, the datahandle_new() call can serve as a cheap auth for cachehit
requests, thus removing the need for our custom auth-calls in Go.

Secondly, we can now introduce a probing call to the C-API that gives
*us* basic information about the file (such as it's shape) before we do
the actual reads, without having to worry too much about performance of
an extra round-trip. This will again open for a range of
possibilities:

1) Re-implement the memory model for data buffers

In the current model, we allocate data buffers in C++ that OpenVDS
writes into and we then expose to Go. We manged without a single copy
in C++, but as soon as we get back to Go, we have to make a copy into a
Go []byte object.

With the VDS shape known up front, we can turn this model on it's head
and allocate buffers in Go upfront and let OpenVDS write straight into
these, saving as a copy.

2) Provides enough information such that Go can be the driver for a
   distributed system

At some point we probably want to split requests and distribute
sub-requests on multiple machines to improve performance. It makes
perfect sense that Go is the driver for such distribution, and this will
provide enough VDS information that go can efficiently split requests.

3) Move all input validations to Go.

Currently most input validations are handle by Go, but some are
dependent on information in the VDS file and those are thus performed in
Go. It will now be cheaper to move these to Go, which makes sense from
an organizational perspective, but will also allow for better error
signaling (as the current error handling through C is limited).

Ofc, all of this is future work.