Add mapping from Go errors to HTTP status codes
As we are starting to get custom error all the way from C++, through
status codes in CGO and then custom Go errors, the last step is to map
the custom Go errors into HTTP status codes that can be returned to the
user.

This mapping should be centralized and not something every gin handler
manages itself, to avoid both codebloat and inconsistencies.