Add an error handling helper for endpoint handlers
Errorhandling in Go is quite verbose. This is particularly evident in our
endpoint handler functions which mostly constist of a linear series of function
call. Each function call is accompanies by 4 lines or error handling. The same
4 lines for every function call.

This new helper addressed this with an interface that allows for short and
concise oneline error handling.