Make C.response an output parameter
The current API design, and the response struct in partular is poor.
response is both a buffer protocol and an error signaler. A better
design, and a common one in C API's is to reserve the return value for
integer status codes and write results into output parameters. This
commit start with the later - freeing up the return value for status
codes.