Formalize the concept of a vertical window
A vertical window defines a window around some arbitrary reference
point. It's used for attribute requests where the caller requests to
calculate attributes along a horizon, including some number of sample
above and below the horizon position - i.e. a vertical window.

This doesn't really add that much to the current code, but it will be
essential once we start supporting subsampling.

The VerticalWindow class is purly a definition of a window's shape. It
does not encapulate any data, nor even a reference point for which the
window spans around. This makes the window definition valid for all
reference points (horizon values). This choice is rooted in performance,
as it means we don't need to construct a window for every depth
(reference point) in the input horizon.