Add support for custom stepsize in /attributes
A common part of calculating attributes around a horizon is to subsample
the traces in the requested vertical window (above/below). This let's
the attribute capure peaks that are not present in the course sampling
of the original seismic cube.

The easy approach
-----------------

The easy way to implement this would be to offload all the work to
OpenVDS. I.e. for every point in the input surface, generate all
vertical points. E.g. using a python range for illustration:

        range(point - above, point + below, stepsize)

Unfortuneally there are 2 major problems with this - memory consumption
and performance.

The memory consumption stems from the fact that when asking openvds for
a specific point, the input is an array of 6 elements, each specifying
the linenumber in one dimension. This quickly becomes enormous. Let's
look at an example:

Request to the API:

        above = 20
        below = 20
        samplestep = 0.1
        surface.shape = (400, 700)

will yeidl a request to OpenVDS of size:

        (400 x 700) x 401 x 24B = 2.694.720.000B = 2.695GB

where (400 x 700) is the number of points in the horizontal plane. 401
is the number of vertical samples to fetch for every horizontal point
(20 / 0.1 + 1 + 20 / 0.1) and 24B is the size of the 6 element float
array needed to represent a single point. That's a lot of data for a
relatively small surface. OpenVDS will even make deep copies of this
data, so the total memory requirement becomes huge.

The large request size to OpenVDS is also partially responsible for the
poor performance that this approach yields (the above query takes around
27 seconds on my machine). Another reason this is slow, is that OpenVDS
threats every point individually and must interpolate it in 3 dimensions
(or potentially even 6).

The implemented approach
------------------------

High level psudo-code of the implemented approach:

    data = fetch_data()

    for window in data:
        window = resample(window)
        for attr in requested_attribute:
                attr.calculate(window)

First, data is fetch from the OpenVDS at a samplerate equal to that in
the VDS file itself. To continue the example, that is every 4ms, which
yields a vertical window of 11 samples (we also needs a few extra
samples at each end for better interpolation, so in practise this window
is 15 samples). The request to OpenVDS now has 40 times fewer points
than in the first apporach (input size is now 67MB vs 2.695 GB).

Then we iterate over all the vertical windows (which are contigious in
memory), resample the window, and calculate all the attribute for that
window.

This approach means we never need to have more than one upsampled window
in memory at the time and since we know samples are all above each
other, we can do 1D interpolation. These two fact makes reduces the
total request time from 27s to around 1s, as well as having a much
smaller peak memory footprint.

Some care must be taken with this approach. Most notably we need to make
sure to fetch samples at exact voxel positions in the vds to avoid
double interpolation. The fetched window and the resampled windows then
need to match up so that we interpolate the right window. However, as
we've seen this added complexity is well worth it.