Compute attributes concurrently
Attribute calculations are very parallelizable as vertical windows are
independent of each other. The concurrency is being driven by go and
goroutines by splitting the set off vertical windows into chunks.

There is more work to be done to find the optimal concurrency level, and
perhaps make it customizable on server startup.