Upgrade c++11 -> c++17
We will soon have boost's makima interpolation as a dependancy and it
requires c++14. If we are upgrading, let's go straight for c++17 to get
some nice new features as-well. There is really no need for us to strive
for portabillity to older systems so c++17 should be safe to use.