Use own class to represent openvds coordinate
There were two reasons to introduce that class.
- It is not straightforward to see what dimension one refers when
DoubleVector3 is addressed by index
- Sharp edge between Request- functions and IJKCoordinateTransformer
makes it difficult to keep track what current DoubleVector3 represents -
is that data adjusted for use in openvds Request- functions or not.

As class is actually a light wrapper over the object we already create,
it should not affect the performance.