Support signed URLs in RequestedResource.Vds
Added NormalizeConnection function which will split the vds URL on the
bases of ‘?’. We are assuming that the first parameter provided by the
user will always be the SAS token. If the user provides any other
parameter with the URL the SAS token the Openvds/Azure may send
authentication failure or return some incorrect data values.