Return geospacial information in slice metadata
Geospacial information is often integral in order to relate seismic to
other data sources. Currently we provide the bounding box of the entire
cube through /metadata. However, when fetching a (vertical) slice the
user would have to manually do a coordinate mapping to get the
geospacial position of the slice. This commit adds this information to
the slice metadata. For horizontal slices, this is just the bounding
box, while for vertical slices this is essentially a linestring.

There are many ways to implement this, and I landed on basing it on the
SubVolume bounds. This has 2 benefits. Firstly, it ensures consistency
between fetch_slice and fetch_slice_metadata as they both are based on
SubVolume's logic. And secondly that makes it future-proof if we ever
deside to implement sub-slicing