Use Azure Key Vaults to obtain allowlist in Radix
It is troublesome to find a good place to keep allowed storage accounts
list. Thus we want to change to a better solution than keeping it
locally in team notes.

Keeping the list as env variable in Radix (not as secret) would make it
visible and more easily changable, but Radix is not a permanent storage.
So if our app gets deleted, the list goes with it.

Other suggestions like finding out if storage account belongs to
specified tenant seem not be an option for us as they require far wider
privileges than the ones we have.

Keeping allowlist as a secret in Key Vault[1] assures list is
persistent. We still have to update the list ourselves (unless we create
a separate solution on top), but the team can see previous versions of
the secret and thus create a new version of the same secret with updated
list.

Note that on version change the secret is updated by Radix, but because
we set the variables up on the start of server, our running server does
not benefit from that update and has to be restarted manually.
But as we do not expect frequent changes to the allowlist, the behavior
is good enough for now.

[1] https://radix.equinor.com/guides/azure-key-vaults/