Fetch horizon in goroutines
According to investigation, using goroutines reduces horizon fetch time
up
to 30%.

Note that in future we will probably split this task into several
goroutines that create their own datahandles/surfaces anyway. So this is
probably a half-way
solution.

Note that chunkSize and maxConcurrentGoroutines might not be optimal.
They are based on heuristics.