Combine surface body parameters into entity
Gather all surface parameters under one name.

As we are planning to add a possibility for users to calculate data
between two surfaces, users will have to provide two sets of the same
parameters for those surfaces. It makes more sense to combine them into
one entity (every with own plane definition, horizon values and fill
value).
As this would be done for a two-surfaces case, it makes sense to do the
same for one-surface case too.

Note that this is a breaking change, but due to us not yet having an
official version and having small enough user-base, we do it without any
additional precautions.