Don't use float/double as iteration var raw loops
Using floating point as an iteration variable in a raw for loop is prone
to floating point rounding errors and can cause the loop to lose one
iteration. We actually observed this for this particular loop, and the
consequence is that we add too few voxels. The list of voxels becomes
left-shifted by one each time this happens, causing us to return
incorrect data to the user.